Added Create Reel video merge

+++ src/pages/AIStudio.tsx
@@ -564,7 +564,7 @@ const AIStudio = () => {
       setMergeProgress(90);
 
       const outputData = await ffmpeg.readFile('output.mp4');
-      const blob = new Blob([outputData], { type: 'video/mp4' });
+      const blob = new Blob([outputData as Uint8Array], { type: 'video/mp4' });
       const url = URL.createObjectURL(blob);
       setMergedReelUrl(url);
       setShowReelDialog(true);
